Реализовал удаление задачи с подтверждением

+++ webapp/views.py
@@ -38,7 +38,9 @@ def task_detail(request, pk):
     return render(request, 'task_detail.html', {'task': task})
 
 
-def task_delete(request, task_id):
-    task = Task.objects.get(id=task_id)
-    task.delete()
-    return redirect('task_list')
+def task_delete(request, pk):
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'POST':
+        task.delete()
+        return redirect('task_list')
+    return render(request, 'task_delete.html', {'task': task})
